Set the type inference in a module
