+++ src/app/pages/ArticlePage/ArticlePage.jsx
@@ -2,7 +2,6 @@ import React, { useContext } from 'react';
 import path from 'ramda/src/path';
 import pathOr from 'ramda/src/pathOr';
 import propEq from 'ramda/src/propEq';
-import isEmpty from 'ramda/src/isEmpty';
 import styled from '@emotion/styled';
 import { string, node } from 'prop-types';
 import useToggle from '#hooks/useToggle';
@@ -66,6 +65,7 @@ import RelatedTopics from '#containers/RelatedTopics';
 import NielsenAnalytics from '#containers/NielsenAnalytics';
 import ScrollablePromo from '#components/ScrollablePromo';
 import bylineExtractor from './utilities/bylineExtractor';
+import categoryName from './utilities/categoryName';
 import Byline from './Byline';
 import getAuthorTwitterHandle from './getAuthorTwitterHandle';
 import { ServiceContext } from '../../contexts/ServiceContext';
@@ -119,7 +119,8 @@ const MpuContainer = styled(AdContainer)`
 
 const ArticlePage = ({ pageData, mostReadEndpointOverride }) => {
   const { isAmp, showAdsBasedOnLocation } = useContext(RequestContext);
-  const { articleAuthor, isTrustProjectParticipant, showRelatedTopics } = useContext(ServiceContext);
+  const { articleAuthor, isTrustProjectParticipant, showRelatedTopics } =
+    useContext(ServiceContext);
   const { enabled: preloadLeadImageToggle } = useToggle('preloadLeadImage');
   const { enabled: adsEnabled } = useToggle('ads');
   const recommendationsData = path(['recommendations'], pageData);
@@ -152,39 +153,10 @@ const ArticlePage = ({ pageData, mostReadEndpointOverride }) => {
     ? getAuthorTwitterHandle(blocks)
     : null;
 
-  const categoryType = taggings => {
-    const supportedPredicate = taggings.filter(
-      tagging =>
-        tagging.predicate ===
-        'http://www.bbc.co.uk/ontologies/creativework/format',
-    );
-
-    if (!isTrustProjectParticipant || isEmpty(supportedPredicate)) {
-      return 'Article';
-    }
-
-    const predicateFormats = pathOr(
-      [],
-      ['metadata', 'passport', 'predicates', 'formats'],
-      pageData,
-    );
-
-    const typeMapping = {
-      'http://www.bbc.co.uk/things/170f311e-fd87-4255-85db-8b4aee12594d#id': 'AnalysisNewsArticle',
-      'http://www.bbc.co.uk/things/bb3ead7c-eb94-453b-b9de-34a59aba2fbf#id': 'BackgroundNewsArticle',
-      'http://www.bbc.co.uk/things/654e1609-7a9f-470a-bd87-fb8e8144c374#id': 'OpinionNewsArticle',
-      'http://www.bbc.co.uk/things/46c0517d-9927-4d1a-9954-8c63a3f7a888#id': 'ReportageNewsArticle',
-      'http://www.bbc.co.uk/things/8d1509ef-08ef-42bd-b831-82504eed9b8e#id': 'ReportageNewsArticle',
-      'http://www.bbc.co.uk/things/3ded0816-13bf-4f1b-90d6-88f1686803d0#id': 'ReviewNewsArticle',
-      'http://www.bbc.co.uk/things/08ac2b5c-889b-4597-821b-de3ed1d0ff15#id': 'ReportageNewsArticle',
-      'http://www.bbc.co.uk/things/a057c92f-2d7e-4997-8310-80ca621d19fa#id': 'ReportageNewsArticle',
-    };
-
-    const articleType = typeMapping[path([0, 'value'], predicateFormats)];
-
-    return articleType ?? 'Article';
-  };
-  const taggings = pathOr([], ['metadata', 'passport', 'taggings'], pageData);
+  const {
+    taggings,
+    predicates: { formats },
+  } = path(['metadata', 'passport'], pageData);
 
   const componentsToRender = {
     visuallyHiddenHeadline,
@@ -284,7 +256,7 @@ const ArticlePage = ({ pageData, mostReadEndpointOverride }) => {
       <LinkedData
         showAuthor
         bylineLinkedData={bylineLinkedData}
-        type={categoryType(taggings)}
+        type={categoryName(taggings, formats, isTrustProjectParticipant)}
         seoTitle={headline}
         headline={headline}
         datePublished={firstPublished}
